+++ src/SVG/index.ts
@@ -1,5 +1,3 @@
-export * from './MarkerElement';
-export * from './PolygonElement';
 export * from './GroupElement';
 export * from './SVG';
 export * from './PathBuilder';