+++ src/style/LinkStyle.ts
@@ -1,0 +1,9 @@
+
+export interface LinkStyleFactory {
+    [key:string]: LinkStyle;
+}
+
+export interface LinkStyle {
+    applySVGPathStyle(path: SVGPathElement): void;
+    getLinkStyleName(): string;    
+}