+++ src/style/MarkerElementStyle.ts
@@ -1,5 +1,7 @@
 export interface MarkerElementStyle {
     id: string;
     fill: string;
-    class: string;    
+    class: string;
+    strokeWidth: string;
+    stroke: string;    
 }