+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog for @codeffekt/ce-canvas-nodes
 
+## [0.0.7]
+ - add link style attribute
+ - add a set of markers
+ - add svg drawing tests 
+
 ## [0.0.6]: correct drag & drop 
 
 ## [0.0.5]: add x,y elt attributes for nodes